+++ app/src/androidTest/java/woowacourse/movie/presentation/view/MovieDetailActivityTest.kt
@@ -16,6 +16,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
+import woowacourse.movie.presentation.ui.detail.MovieDetailActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieDetailActivityTest {
