+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.ReservationResultActivity"
+            android:name=".presentation.ui.reservation.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.MovieDetailActivity"
+            android:name=".presentation.ui.detail.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
