+++ app/src/main/java/woowacourse/movie/presentation/ui/main/adapter/MovieListAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.adapter
+package woowacourse.movie.presentation.ui.main.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
